refactor: separate super blink for led color cycle
