+++ src/common/common.h
@@ -1,0 +1,3 @@
+#define COMMON_DELAY 200
+
+void delay(int milliseconds);
