+++ src/led_color_cycle_example/led_color_cycle.h
@@ -1,0 +1,2 @@
+#pragma once
+int run_led_color_cycle();
