+++ src/main.c
@@ -1,0 +1,9 @@
+#include "led_color_cycle_example/led_color_cycle.h"
+#include "super_blink_example/super_blink.h"
+#include <stdio.h>
+
+int main()
+{
+    run_led_color_cycle();
+    // run_super_blink();
+}
